Separate the waker executor into its own crate
