+++ applications/test_multitask/src/lib.rs
@@ -7,12 +7,12 @@ extern crate alloc;
 
 use alloc::string::String;
 use alloc::vec::Vec;
-use simple_executor::SimpleExecutor;
+use waker_executor::WakerExecutor;
 use task_async::TaskAsync;
 use timer_future::TimerFuture;
 
 pub fn main(_args: Vec<String>) -> isize {
-    let mut executor = SimpleExecutor::new();
+    let mut executor = WakerExecutor::new();
     executor.spawn(TaskAsync::new(async {
         for _ in 0..3 {
             println!("howdy 1!");
@@ -28,7 +28,10 @@ pub fn main(_args: Vec<String>) -> isize {
             println!("done 2!")
         }
     }));
-    executor.run();
+    
+    if let Err(_e) = executor.run() {
+        println!("executor had an error")
+    }
 
     0
 }
