+++ kernel/simple_executor/Cargo.toml
@@ -4,19 +4,8 @@ name = "simple_executor"
 description = "A simple executor for async tasks"
 version = "0.1.0"
 
-[dependencies]
-x86_64 = "0.14.8"
-
-[dependencies.crossbeam-queue]
-version = "0.3.8"
-default-features = false
-features = ["alloc"]
-
 [dependencies.task_async]
 path = "../task_async"
-
-[dependencies.terminal_print]
-path = "../../kernel/terminal_print"
 
 [lib]
 crate-type = ["rlib"]
