+++ kernel/waker_executor/src/lib.rs
@@ -1,0 +1,125 @@
+#![no_std]
+
+#[macro_use]
+extern crate terminal_print;
+extern crate alloc;
+extern crate crossbeam_queue;
+extern crate task_async;
+extern crate x86_64;
+
+use alloc::{collections::BTreeMap, sync::Arc, task::Wake};
+use core::task::{Context, Poll, Waker};
+use crossbeam_queue::ArrayQueue;
+use task_async::{TaskAsync, TaskId};
+
+pub struct WakerExecutor {
+    // tasks are accessed by TaskId in the map
+    tasks: BTreeMap<TaskId, TaskAsync>,
+    // wrapped into Arc type that implements reference counting
+    task_queue: Arc<ArrayQueue<TaskId>>,
+    waker_cache: BTreeMap<TaskId, Waker>,
+}
+
+impl WakerExecutor {
+    pub fn new() -> WakerExecutor {
+        WakerExecutor {
+            tasks: BTreeMap::new(),
+            task_queue: Arc::new(ArrayQueue::new(100)),
+            waker_cache: BTreeMap::new(),
+        }
+    }
+
+    // push the corresponding async task in the queue when spawned
+    pub fn spawn(&mut self, task: TaskAsync) {
+        let task_id = task.id;
+
+        // check if task ID exists and if the queue is full
+        if self.tasks.insert(task.id, task).is_some() {
+            panic!("task with the same ID already in tasks");
+        }
+        self.task_queue.push(task_id).expect("queue full");
+    }
+
+    pub fn run(&mut self) -> Result<(), ()> {
+        loop {
+            if self.tasks.is_empty() {
+                return Ok(());
+            }
+            self.run_ready_tasks();
+            self.sleep_if_idle();
+        }
+    }
+
+    fn run_ready_tasks(&mut self) {
+        // destructure `self` to avoid borrow checker errors
+        let Self {
+            tasks,
+            task_queue,
+            waker_cache,
+        } = self;
+
+        // loop over all tasks in the task queue
+        while let Some(task_id) = task_queue.pop() {
+            let task = match tasks.get_mut(&task_id) {
+                Some(task) => {
+                    println!("Task ID: {:?}", task_id);
+                    task
+                },
+                None => continue, // task doesn't exist anymore
+            };
+
+            let waker = waker_cache
+                .entry(task_id)
+                .or_insert_with(|| TaskWaker::new(task_id, task_queue.clone()));
+            let mut context = Context::from_waker(waker);
+            match task.poll(&mut context) {
+                Poll::Ready(()) => {
+                    // task is done so remove it and its cached waker
+                    tasks.remove(&task_id);
+                    waker_cache.remove(&task_id);
+                }
+                Poll::Pending => {}
+            }
+        }
+    }
+
+    fn sleep_if_idle(&self) {
+        use x86_64::instructions::interrupts::{self, enable_and_hlt};
+
+        interrupts::disable();
+        if self.task_queue.is_empty() {
+            enable_and_hlt();
+        } else {
+            interrupts::enable();
+        }
+    }
+}
+
+// waker implementation
+struct TaskWaker {
+    task_id: TaskId,
+    task_queue: Arc<ArrayQueue<TaskId>>,
+}
+
+impl TaskWaker {
+    fn new(task_id: TaskId, task_queue: Arc<ArrayQueue<TaskId>>) -> Waker {
+        Waker::from(Arc::new(TaskWaker {
+            task_id,
+            task_queue,
+        }))
+    }
+
+    fn wake_task(&self) {
+        self.task_queue.push(self.task_id).expect("task_queue full");
+    }
+}
+
+impl Wake for TaskWaker {
+    fn wake(self: Arc<Self>) {
+        self.wake_task();
+    }
+
+    fn wake_by_ref(self: &Arc<Self>) {
+        self.wake_task();
+    }
+}
